### 2025-04-07 2b5b6df
binary initialization fix
