### 2025-04-11 #8024 @crvernon
:wave: @povilaskarvelis - I am going to run this by our larger editorial board for review to ensure it meets our requirements for substantial scholarly effort and research software requirements.  I'll get back in touch with you ASAP with the outcome.  Thanks!
