### 2025-05-01 #8024 @crvernon
:wave: @jromanowska - Are you still able to take this one on?  I am fine if you are not able to start until mid-May as you stated since we have a longer backlog of papers currently.  If you can, we will go ahead and assign this one to you.  Hope you have a great day!
