### 2025-05-15 97b0077
smol readme edits
