### 2025-05-25 #8024 @IndrajeetPatil
@jromanowska Hi! Unfortunately, I wouldn't be able to review this, but I could suggest @mattansb, who is the author of effectsize package and has also published with [JOSS](https://joss.theoj.org/papers/10.21105/joss.02815), as an alternative :)
