### 2025-05-27 58042ff
readme smol rewording
