### 2025-06-02 #8024 @jromanowska
> Hi, [@povilaskarvelis](https://github.com/povilaskarvelis) 👋 I will take care of the process from now on. I will try to find reviewers, but first could you fix the DOIs in the paper, as marked above by editorialbot?

@povilaskarvelis, please fix the DOIs.

@bede , @lauragarrison87, we will wait only for the author to fix that last thing and then, I will start the official review.
