### 2025-06-03 171ea61
added dois to bibtex
