+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={175--191},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.3758/bf03193146}
 }
 
 @article{maher2013other,
@@ -17,7 +18,8 @@
   number={3},
   pages={345--351},
   year={2013},
-  publisher={American Society for Cell Biology}
+  publisher={American Society for Cell Biology},
+  doi={10.1187/cbe.13-04-0082}
 }
 
 @article{karvelis2023individual,
@@ -27,7 +29,8 @@
   volume={148},
   pages={105137},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.31234/osf.io/bvjzn}
 }
 
 
@@ -35,7 +38,8 @@
   title={Clarifying the reliability paradox: poor test-retest reliability attenuates group differences},
   author={Karvelis, Povilas and Diaconescu, Andreea O},
   year={2025},
-  publisher={PsyArXiv}
+  publisher={PsyArXiv},
+  doi={10.31234/osf.io/z4yqe_v1}
 }
 
 @article{abi2016search,
@@ -46,7 +50,8 @@
   number={11},
   pages={1248--1255},
   year={2016},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/nm.4190}
 }
 
 @article{large2018role,
@@ -68,7 +73,8 @@
   number={2},
   pages={236--262},
   year={2023},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wps.21078}
 }
 
 @article{hedge2018reliability,
@@ -78,7 +84,8 @@
   volume={50},
   pages={1166--1186},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.3758/s13428-017-0935-1}
 }
 
 @article{funder2019evaluating,
@@ -100,7 +107,8 @@
   number={3},
   pages={1068--1072},
   year={2018},
-  publisher={LWW}
+  publisher={LWW},
+  doi={10.1213/ane.0000000000002798}
 }
 
 @article{monsarrat2018intriguing,
@@ -111,11 +119,12 @@
   number={1},
   pages={gix121},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/gigascience/gix121}
 }
 
 @misc{wasserstein2019moving,
-  title={Moving to a world beyond “p< 0.05”},
+  title={Moving to a world beyond "p< 0.05"},
   author={Wasserstein, Ronald L and Schirm, Allen L and Lazar, Nicole A},
   journal={The American Statistician},
   volume={73},
@@ -133,7 +142,8 @@
   number={8},
   pages={855--859},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jclinepi.2015.02.010}
 }
 
 @article{meehl1990summaries,
@@ -144,7 +154,8 @@
   number={1},
   pages={195--244},
   year={1990},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={10.2466/pr0.66.1.195-244}
 }
 
 @article{gignac2016effect,
@@ -154,7 +165,8 @@
   volume={102},
   pages={74--78},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.paid.2016.06.069}
   
 }
 
