### 2025-06-03 5c306ea
added dois to bibtex
