+++ paper/paper.bib
@@ -39,7 +39,7 @@
   author={Karvelis, Povilas and Diaconescu, Andreea O},
   year={2025},
   publisher={PsyArXiv},
-  doi={10.31234/osf.io/z4yqe_v1}
+  doi={10.31234/osf.io/z4yqe_v4}
 }
 
 @article{abi2016search,
