### 2025-06-03 d15204b
added remaining dois to bibtex
