+++ paper/paper.bib
@@ -62,7 +62,8 @@
   number={3},
   pages={197--205},
   year={2018},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.31887/DCNS.2018.20.3/mlarge}
 }
 
 @article{abi2023candidate,
@@ -96,7 +97,8 @@
   number={2},
   pages={156--168},
   year={2019},
-  publisher={Sage Publications Sage CA: Los Angeles, CA}
+  publisher={Sage Publications Sage CA: Los Angeles, CA},
+  doi={10.1177/2515245919847202}
 }
 
 @article{schober2018statistical,
@@ -131,7 +133,8 @@
   number={sup1},
   pages={1--19},
   year={2019},
-  publisher={Taylor \& Francis}
+  publisher={Taylor & Francis},
+  doi={10.1080/00031305.2019.1583913}
 }
 
 @article{ozenne2015precision,
@@ -178,7 +181,8 @@
   number={2},
   pages={129--133},
   year={2016},
-  publisher={Taylor \& Francis}
+  publisher={Taylor & Francis},
+  doi={10.1080/00031305.2016.1154108}
 }
 
 @article{d3js,
