### 2025-06-03 #8334 @editorialbot
[REVIEW]: E2P Simulator: An interactive tool for translating effect sizes into predictive utility
**Submitting author:** @povilaskarvelis (<a href="http://orcid.org/0000-0001-7469-5624">Povilas Karvelis</a>)
**Repository:** https://github.com/povilaskarvelis/e2p-simulator
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.2.0
