+++ js/binary.js
@@ -16,12 +16,16 @@ const PLOT_CONFIG = {
 const SELECTORS = {
     overlapPlot: 'overlap-plot',
     rocPlot: 'roc-plot',
-    prPlot: 'pr-plot'
+    prPlot: 'pr-plot',
+    dcaPlot: 'dca-plot'
 };
 
 // State variables
 let thresholdValue = 0;
+let dcaThresholdMin = 0.05; // 5% default minimum threshold
+let dcaThresholdMax = 0.25; // 25% default maximum threshold
 let rocInitialized = false;
+let dcaInitialized = false;
 let currentView = "observed";
 let xScale, yScale;
 let plotGroup;
@@ -495,6 +499,9 @@ function plotROC(d) {
             displayModeBar: false
         };
 
+        // Plot DCA
+        plotDCA(d, baseRate, sigma1, sigma2);
+
         if (!rocInitialized) {
             Plotly.newPlot(SELECTORS.rocPlot, [rocTrace, rocThresholdMarker], rocLayout, config);
             Plotly.newPlot(SELECTORS.prPlot, [prTrace, prThresholdMarker], prLayout, config);
@@ -506,6 +513,540 @@ function plotROC(d) {
     } catch (error) {
         console.error("Error plotting ROC curve:", error);
     }
+}
+
+function plotDCA(d, baseRate, sigma1, sigma2) {
+    try {
+        // Get metrics at current threshold slider position
+        const currentMetrics = computeMetricsForBinaryDistributions(d, baseRate, sigma1, sigma2, thresholdValue);
+        const sensitivity = currentMetrics.sensitivity;
+        const specificity = currentMetrics.specificity;
+        
+        // Calculate net benefit across threshold probabilities using current model performance
+        const ptMin = 0.001;
+        const ptMax = 0.999;
+        const step = 0.001;
+        
+        const thresholdProbs = [];
+        const netBenefits = [];
+        const treatAllBenefits = [];
+        const treatNoneBenefits = [];
+        
+        for (let pt = ptMin; pt <= ptMax; pt += step) {
+            // Use current threshold's sensitivity/specificity for all threshold probabilities
+            const odds = pt / (1 - pt);
+            
+            const netBenefit = (sensitivity * baseRate) - ((1 - specificity) * (1 - baseRate) * odds);
+            
+            // Calculate treat all strategy: treat everyone regardless of test result
+            // TP = all cases, FP = all controls
+            const treatAllBenefit = baseRate - ((1 - baseRate) * odds);
+            
+            thresholdProbs.push(pt);
+            netBenefits.push(netBenefit); // Allow negative net benefit
+            treatAllBenefits.push(treatAllBenefit); // Allow negative net benefit for treat all
+            treatNoneBenefits.push(0); // Treat none = 0
+        }
+        
+        
+        // Create separate shaded areas for positive and negative net benefits
+        const positiveX = [];
+        const positiveY = [];
+        const negativeX = [];
+        const negativeY = [];
+        
+        for (let i = 0; i < thresholdProbs.length; i++) {
+            const pt = thresholdProbs[i];
+            if (pt >= dcaThresholdMin && pt <= dcaThresholdMax) {
+                const netBenefit = netBenefits[i];
+                
+                if (netBenefit >= 0) {
+                    positiveX.push(pt);
+                    positiveY.push(netBenefit);
+                } else {
+                    negativeX.push(pt);
+                    negativeY.push(netBenefit);
+                }
+            }
+        }
+        
+        // Create traces
+        // Calculate A-NBC (Area Under Net Benefit Curve) using trapezoidal integration
+        let anbc = 0;
+        const rangeStart = dcaThresholdMin;
[…]
