+++ js/continuous.js
@@ -27,12 +27,14 @@ const SELECTORS = {
     distributionPlotTrue: 'distribution-plot-true-cont',
     distributionPlotObserved: 'distribution-plot-observed-cont',
     rocPlot: 'roc-plot-cont',
-    prPlot: 'pr-plot-cont'
+    prPlot: 'pr-plot-cont',
+    dcaPlot: 'dca-plot-cont'
 };
 
 // State variables
 let thresholdValue = 0;
 let rocInitialized = false;
+let dcaInitialized = false;
 let currentView = "observed";
 let trueLabeledData = [];
 let observedLabeledData = [];
@@ -40,6 +42,10 @@ let currentLabeledData = [];
 let trueMetrics = {};
 let observedMetrics = {};
 let xScale, yScale;
+
+// DCA threshold range variables
+let dcaThresholdMin = 0.05; // 5%
+let dcaThresholdMax = 0.25; // 25%
 
 // Utility functions
 function computeObservedR(trueR, reliabilityX, reliabilityY) {
@@ -163,6 +169,9 @@ function cleanupContinuous() {
     // Reset state and clean up plots (existing cleanup)
     Plotly.purge(SELECTORS.rocPlot);
     Plotly.purge(SELECTORS.prPlot); 
+    if (document.getElementById(SELECTORS.dcaPlot)) {
+        Plotly.purge(SELECTORS.dcaPlot);
+    }
     // Add purging/cleanup for D3 plots if necessary
     d3.select(`#${SELECTORS.scatterPlotTrue}`).selectAll("*").remove();
     d3.select(`#${SELECTORS.scatterPlotObserved}`).selectAll("*").remove();
@@ -172,11 +181,14 @@ function cleanupContinuous() {
     // Reset global vars if needed
     thresholdValue = 0;
     rocInitialized = false;
+    dcaInitialized = false;
     trueMetrics = {};
     observedMetrics = {};
     currentView = "observed";
     trueLabeledData = [];
     observedLabeledData = [];
+    dcaThresholdMin = 0.05;
+    dcaThresholdMax = 0.25;
 }
 
 // Drawing functions
@@ -741,6 +753,520 @@ function plotROC() {
     } else {
         Plotly.react(SELECTORS.rocPlot, [rocTrace, thresholdMarker], rocLayout, config);
         Plotly.react(SELECTORS.prPlot, [prTrace, prThresholdMarker], prLayout, config);
+    }
+    
+    // Plot DCA (only if container exists)
+    if (document.getElementById(SELECTORS.dcaPlot)) {
+        plotDCA();
+    }
+}
+
+function plotDCA() {
+    if (!currentLabeledData || currentLabeledData.length === 0) {
+        console.warn("plotDCA called with no currentLabeledData.");
+        return;
+    }
+    
+    try {
+        // Get base rate from the data
+        const baseRate = currentLabeledData.filter(d => d.trueClass === 1).length / currentLabeledData.length;
+        
+        // Get metrics at current threshold position
+        const currentMetrics = computePredictiveMetrics(thresholdValue, currentLabeledData);
+        const sensitivity = currentMetrics.sensitivity;
+        const specificity = currentMetrics.specificity;
+        
+        // Calculate net benefit across threshold probabilities using current model performance
+        const ptMin = 0.001;
+        const ptMax = 0.999;
+        const step = 0.001;
+        
+        const thresholdProbs = [];
+        const netBenefits = [];
+        const treatAllBenefits = [];
+        const treatNoneBenefits = [];
+        
+        for (let pt = ptMin; pt <= ptMax; pt += step) {
+            // Use current threshold's sensitivity/specificity for all threshold probabilities
+            const odds = pt / (1 - pt);
+            
[…]
