### 2025-06-08 dca6d47
fixed dca red shading area
