+++ js/continuous.js
@@ -34,7 +34,6 @@ const SELECTORS = {
 // State variables
 let thresholdValue = 0;
 let rocInitialized = false;
-let dcaInitialized = false;
 let currentView = "observed";
 let trueLabeledData = [];
 let observedLabeledData = [];
@@ -42,10 +41,6 @@ let currentLabeledData = [];
 let trueMetrics = {};
 let observedMetrics = {};
 let xScale, yScale;
-
-// DCA threshold range variables
-let dcaThresholdMin = 0.05; // 5%
-let dcaThresholdMax = 0.25; // 25%
 
 // Utility functions
 function computeObservedR(trueR, reliabilityX, reliabilityY) {
@@ -169,9 +164,12 @@ function cleanupContinuous() {
     // Reset state and clean up plots (existing cleanup)
     Plotly.purge(SELECTORS.rocPlot);
     Plotly.purge(SELECTORS.prPlot); 
-    if (document.getElementById(SELECTORS.dcaPlot)) {
-        Plotly.purge(SELECTORS.dcaPlot);
+    
+    // Cleanup DCA module
+    if (typeof DCAModule !== 'undefined') {
+        DCAModule.cleanup('continuous');
     }
+    
     // Add purging/cleanup for D3 plots if necessary
     d3.select(`#${SELECTORS.scatterPlotTrue}`).selectAll("*").remove();
     d3.select(`#${SELECTORS.scatterPlotObserved}`).selectAll("*").remove();
@@ -181,14 +179,11 @@ function cleanupContinuous() {
     // Reset global vars if needed
     thresholdValue = 0;
     rocInitialized = false;
-    dcaInitialized = false;
     trueMetrics = {};
     observedMetrics = {};
     currentView = "observed";
     trueLabeledData = [];
     observedLabeledData = [];
-    dcaThresholdMin = 0.05;
-    dcaThresholdMax = 0.25;
 }
 
 // Drawing functions
@@ -755,520 +750,23 @@ function plotROC() {
         Plotly.react(SELECTORS.prPlot, [prTrace, prThresholdMarker], prLayout, config);
     }
     
-    // Plot DCA (only if container exists)
-    if (document.getElementById(SELECTORS.dcaPlot)) {
-        plotDCA();
-    }
-}
-
-function plotDCA() {
-    if (!currentLabeledData || currentLabeledData.length === 0) {
-        console.warn("plotDCA called with no currentLabeledData.");
-        return;
-    }
-    
-    try {
-        // Get base rate from the data
+    // Plot DCA (only if container exists and DCA module is available)
+    if (document.getElementById(SELECTORS.dcaPlot) && typeof DCAModule !== 'undefined') {
+        // Get base rate from current data
         const baseRate = currentLabeledData.filter(d => d.trueClass === 1).length / currentLabeledData.length;
         
         // Get metrics at current threshold position
         const currentMetrics = computePredictiveMetrics(thresholdValue, currentLabeledData);
-        const sensitivity = currentMetrics.sensitivity;
-        const specificity = currentMetrics.specificity;
         
-        // Calculate net benefit across threshold probabilities using current model performance
-        const ptMin = 0.001;
-        const ptMax = 0.999;
-        const step = 0.001;
-        
-        const thresholdProbs = [];
-        const netBenefits = [];
-        const treatAllBenefits = [];
-        const treatNoneBenefits = [];
-        
-        for (let pt = ptMin; pt <= ptMax; pt += step) {
-            // Use current threshold's sensitivity/specificity for all threshold probabilities
-            const odds = pt / (1 - pt);
-            
-            const netBenefit = (sensitivity * baseRate) - ((1 - specificity) * (1 - baseRate) * odds);
-            
[…]
