+++ js/dca.js
@@ -1,0 +1,524 @@
+(function() {
+// DCA Module - Modular Decision Curve Analysis implementation
+const DCAModule = {
+    // State variables for each DCA instance
+    instances: new Map(),
+    
+    // Initialize a DCA instance
+    init: function(instanceId, config) {
+        this.instances.set(instanceId, {
+            plotSelector: config.plotSelector,
+            thresholdMin: config.thresholdMin || 0.05,
+            thresholdMax: config.thresholdMax || 0.25,
+            initialized: false,
+            onThresholdChange: config.onThresholdChange || (() => {})
+        });
+    },
+    
+    // Plot DCA for a given instance
+    plot: function(instanceId, data) {
+        const instance = this.instances.get(instanceId);
+        if (!instance) {
+            console.error(`DCA instance ${instanceId} not found`);
+            return;
+        }
+        
+        try {
+            const {
+                sensitivity,
+                specificity,
+                baseRate
+            } = data;
+            
+            // Calculate net benefit across threshold probabilities using current model performance
+            const ptMin = 0.001;
+            const ptMax = 0.999;
+            const step = 0.001;
+            
+            const thresholdProbs = [];
+            const netBenefits = [];
+            const treatAllBenefits = [];
+            const treatNoneBenefits = [];
+            
+            for (let pt = ptMin; pt <= ptMax; pt += step) {
+                // Use current threshold's sensitivity/specificity for all threshold probabilities
+                const odds = pt / (1 - pt);
+                
+                const netBenefit = (sensitivity * baseRate) - ((1 - specificity) * (1 - baseRate) * odds);
+                
+                // Calculate treat all strategy: treat everyone regardless of test result
+                // TP = all cases, FP = all controls
+                const treatAllBenefit = baseRate - ((1 - baseRate) * odds);
+                
+                thresholdProbs.push(pt);
+                netBenefits.push(netBenefit); // Allow negative net benefit
+                treatAllBenefits.push(treatAllBenefit); // Allow negative net benefit for treat all
+                treatNoneBenefits.push(0); // Treat none = 0
+            }
+            
+            // Calculate A-NBC (Area Under Net Benefit Curve) using trapezoidal integration
+            let anbc = 0;
+            const rangeStart = instance.thresholdMin;
+            const rangeEnd = instance.thresholdMax;
+            
+            for (let i = 1; i < thresholdProbs.length; i++) {
+                const pt1 = thresholdProbs[i-1];
+                const pt2 = thresholdProbs[i];
+                
+                // Only include points within the selected range
+                if (pt1 >= rangeStart && pt2 <= rangeEnd) {
+                    const nb1 = netBenefits[i-1];
+                    const nb2 = netBenefits[i];
+                    anbc += (pt2 - pt1) * (nb1 + nb2) / 2; // Trapezoidal rule
+                }
+            }
+            
+            // Create separate shaded areas for positive net benefits
+            const positiveX = [];
+            const positiveY = [];
+            
+            for (let i = 0; i < thresholdProbs.length; i++) {
+                const pt = thresholdProbs[i];
+                if (pt >= instance.thresholdMin && pt <= instance.thresholdMax) {
+                    const netBenefit = netBenefits[i];
+                    
+                    if (netBenefit >= 0) {
+                        positiveX.push(pt);
+                        positiveY.push(netBenefit);
+                    }
+                    // Note: we don't use negativeX/Y here anymore since we handle it separately below
+                }
+            }
+            
+            // Create traces
+            const dcaTrace = {
+                x: thresholdProbs,
+                y: netBenefits,
+                type: "scatter",
+                mode: "lines",
+                name: "Model",
[…]

[…]
