### 2025-06-08 a3fc659
added dca and aucs figs to get started
