+++ js/continuous.js
@@ -744,6 +744,15 @@ function plotROC() {
     if (!rocInitialized) {
         Plotly.newPlot(SELECTORS.rocPlot, [rocTrace, thresholdMarker], rocLayout, config);
         Plotly.newPlot(SELECTORS.prPlot, [prTrace, prThresholdMarker], prLayout, config);
+        
+        // Add click event listeners to navigate to get-started.html sections
+        document.getElementById(SELECTORS.rocPlot).addEventListener('click', () => {
+            window.open('get-started.html#threshold-metrics', '_blank');
+        });
+        document.getElementById(SELECTORS.prPlot).addEventListener('click', () => {
+            window.open('get-started.html#threshold-metrics', '_blank');
+        });
+        
         rocInitialized = true;
     } else {
         Plotly.react(SELECTORS.rocPlot, [rocTrace, thresholdMarker], rocLayout, config);
