+++ js/binary.js
@@ -506,6 +506,15 @@ function plotROC(d) {
         if (!rocInitialized) {
             Plotly.newPlot(SELECTORS.rocPlot, [rocTrace, rocThresholdMarker], rocLayout, config);
             Plotly.newPlot(SELECTORS.prPlot, [prTrace, prThresholdMarker], prLayout, config);
+            
+            // Add click event listeners to navigate to get-started.html sections
+            document.getElementById(SELECTORS.rocPlot).addEventListener('click', () => {
+                window.open('get-started.html#threshold-metrics', '_blank');
+            });
+            document.getElementById(SELECTORS.prPlot).addEventListener('click', () => {
+                window.open('get-started.html#threshold-metrics', '_blank');
+            });
+            
             rocInitialized = true;
         } else {
             Plotly.react(SELECTORS.rocPlot, [rocTrace, rocThresholdMarker], rocLayout, config);
