+++ js/dca.js
@@ -56,8 +56,8 @@ const DCAModule = {
                 treatNoneBenefits.push(0); // Treat none = 0
             }
             
-            // Calculate A-NBC (Area Under Net Benefit Curve) using trapezoidal integration
-            let anbc = 0;
+            // Calculate Delta A-NBC (improvement over best simple strategy)
+            let deltaANBC = 0;
             const rangeStart = instance.thresholdMin;
             const rangeEnd = instance.thresholdMax;
             
@@ -69,24 +69,50 @@ const DCAModule = {
                 if (pt1 >= rangeStart && pt2 <= rangeEnd) {
                     const nb1 = netBenefits[i-1];
                     const nb2 = netBenefits[i];
-                    anbc += (pt2 - pt1) * (nb1 + nb2) / 2; // Trapezoidal rule
+                    const ta1 = treatAllBenefits[i-1];
+                    const ta2 = treatAllBenefits[i];
+                    
+                    // Treat none benefit is always 0
+                    const tn1 = 0;
+                    const tn2 = 0;
+                    
+                    // Find the best simple strategy at each point
+                    const bestSimple1 = Math.max(ta1, tn1); // max(treat all, treat none)
+                    const bestSimple2 = Math.max(ta2, tn2);
+                    
+                    // Calculate improvement over best simple strategy
+                    const improvement1 = nb1 - bestSimple1;
+                    const improvement2 = nb2 - bestSimple2;
+                    
+                    // Integrate the improvement (can be negative)
+                    deltaANBC += (pt2 - pt1) * (improvement1 + improvement2) / 2; // Trapezoidal rule
                 }
             }
             
-            // Create separate shaded areas for positive net benefits
+            // Create separate shaded areas for different strategies
             const positiveX = [];
             const positiveY = [];
+            const positiveBaseline = []; // For bounding the positive area
+            const treatAllShadedX = [];
+            const treatAllShadedY = [];
             
             for (let i = 0; i < thresholdProbs.length; i++) {
                 const pt = thresholdProbs[i];
                 if (pt >= instance.thresholdMin && pt <= instance.thresholdMax) {
                     const netBenefit = netBenefits[i];
+                    const treatAllBenefit = treatAllBenefits[i];
                     
-                    if (netBenefit >= 0) {
+                    // Only include in positive area when model is better than both 0 and treat all
+                    if (netBenefit >= 0 && netBenefit >= treatAllBenefit) {
                         positiveX.push(pt);
                         positiveY.push(netBenefit);
+                        // Use the higher of 0 or treat all as the baseline
+                        positiveBaseline.push(Math.max(0, treatAllBenefit));
                     }
-                    // Note: we don't use negativeX/Y here anymore since we handle it separately below
+                    
+                    // Collect treat all points for shading
+                    treatAllShadedX.push(pt);
+                    treatAllShadedY.push(treatAllBenefit);
                 }
             }
             
@@ -96,21 +122,54 @@ const DCAModule = {
                 y: netBenefits,
                 type: "scatter",
                 mode: "lines",
-                name: "Model",
+                name: "Predictor",
                 line: { color: "black", width: 2 },
+                showlegend: true,
             };
             
-            // Positive shaded area (blue) - always bounded by y=0 at bottom
-            const positiveArea = {
-                x: positiveX,
-                y: positiveY,
+            // Treat all shaded area (gray) - reference strategy
+            const treatAllArea = {
+                x: treatAllShadedX,
+                y: treatAllShadedY,
                 type: "scatter",
                 mode: "none",
                 fill: "tozeroy",
-                fillcolor: "rgba(46, 134, 171, 0.3)",
[…]
