+++ js/binary.js
@@ -431,7 +431,7 @@ function plotROC(d) {
                 y: 0.05,
                 xref: "paper",
                 yref: "paper",
-                text: `AUC: ${auc.toFixed(2)}`,
+                text: `ROC-AUC: ${auc.toFixed(2)}`,
                 showarrow: false,
                 font: { size: 16, color: "black", weight: "bold" },
                 align: "right",
