+++ js/binary.js
@@ -149,6 +149,8 @@ function drawDistributions(d) {
             .attr("opacity", 0.3)
             .attr("d", area);
 
+
+
         // Update legend
         const urlParams = parseURLParams(); // Get URL params
         const label1 = urlParams.label1 || "Group 1"; // Default if not provided
@@ -206,7 +208,7 @@ function drawThreshold(d) {
         const groupEnter = thresholdGroup.enter()
             .append("g")
             .attr("class", "threshold-group")
-            .style("cursor", "pointer");
+            .style("cursor", "ew-resize");
 
         // Merge enter and update selections
         const thresholdMerge = groupEnter.merge(thresholdGroup);
@@ -573,12 +575,14 @@ function updateMetricsFromD(d) {
         const trueLogOddsRatio = StatUtils.dToLogOddsRatio(d);
         const trueR = StatUtils.dToR(d, baseRate);
         const trueEtaSquared = trueR ** 2;
+        const trueCohensU3 = StatUtils.normalCDF(d, 0, 1); // U3 = Φ(d)
 
         // Calculate values for observed d
         const obsOddsRatio = StatUtils.dToOddsRatio(dObs);
         const obsLogOddsRatio = StatUtils.dToLogOddsRatio(dObs);
         const obsR = StatUtils.dToR(dObs, baseRate);
         const obsEtaSquared = obsR ** 2;
+        const obsCohensU3 = StatUtils.normalCDF(dObs, 0, 1); // U3 = Φ(d)
 
         // Update all inputs
         document.getElementById("true-difference-number-bin").value = d.toFixed(2);
@@ -590,12 +594,14 @@ function updateMetricsFromD(d) {
         document.getElementById("true-log-odds-ratio-bin").value = trueLogOddsRatio.toFixed(2);
         document.getElementById("true-pb-r-bin").value = trueR.toFixed(2);
         document.getElementById("true-eta-squared-bin").value = trueEtaSquared.toFixed(2);
+        document.getElementById("true-cohens-u3-bin").value = trueCohensU3.toFixed(2);
 
         // Update observed metrics
         document.getElementById("observed-odds-ratio-bin").value = obsOddsRatio.toFixed(2);
         document.getElementById("observed-log-odds-ratio-bin").value = obsLogOddsRatio.toFixed(2);
         document.getElementById("observed-pb-r-bin").value = obsR.toFixed(2);
         document.getElementById("observed-eta-squared-bin").value = obsEtaSquared.toFixed(2);
+        document.getElementById("observed-cohens-u3-bin").value = obsCohensU3.toFixed(2);
         
         // Update plots
         updatePlots();
@@ -637,6 +643,18 @@ function updateMetricsFromR(r) {
     }
 }
 
+function updateMetricsFromCohensU3(u3) {
+    try {
+        u3 = parseFloat(u3);
+        if (isNaN(u3) || u3 >= 1 || u3 <= 0) return;
+        // Convert U3 to Cohen's d using inverse normal CDF
+        const d = inverseNormalCDF(u3);
+        updateMetricsFromD(d);
+    } catch (error) {
+        console.error("Error updating metrics from Cohen's U3:", error);
+    }
+}
+
 // Function to update plots and metrics based on current state
 function updatePlots() {
     try {
@@ -659,6 +677,23 @@ function updatePlots() {
 }
 
 // Setup event listeners for controls
+// Update highlighting of metric input columns
+function updateMetricsHighlighting(view) {
+    const metricsInputs = document.querySelectorAll('.metrics-input');
+    
+    metricsInputs.forEach(input => {
+        // Remove existing highlighting classes
+        input.classList.remove('selected-true', 'selected-observed');
+        
[…]

[…]
