### 2025-07-09 9bb1fac
title and description edits: reframing the purpose of the simulator
