### 2025-07-09 423eb7b
updated readme and get started to distinguish the utility part more
