### 2025-07-09 46b3315
updated readme and get started with more clarity around utility
