### 2025-07-13 cc35f69
fixed minr bugs in e2e testing
