### 2025-07-13 #8334 @povilaskarvelis
@Annchovy thank you for flagging this - I have now updated the code and the tests run smoothly
