### 2025-07-16 povilaskarvelis/e2p-simulator#3 @Annchovy
Reproducibility of the Quick Start Examples
After launching the examples (Quick Start Examples section) and playing around with some of the parameters (e.g., the observed effect size), I did not manage to reproduce the described results. After changing some of the parameters and then returning them to the suggested values, the resulting metrics were slightly different. For instance, the simulator produced ROC-AUC = 0.72 instead of 0.71 in the example 1 (Diagnostic Prediction), and PR-AUC = 0.33 instead of 0.32 in the example 2 (Treatment Response Prediction).
Screenshots are attached.

Example 1:

Example 2:
