### 2025-07-17 #8334 @Annchovy
Incomplete sentence: https://github.com/povilaskarvelis/e2p-simulator/issues/4#issue-3239102631
