### 2025-07-17 povilaskarvelis/e2p-simulator#5 @Annchovy
Paper
A few comments on the paper to consider:
1) The paper would benefit from an overview of the interface including its components and main functionality.
2) The citation format requires a small refinement: when citing multiple sources, the works should be separated by semicolons (";") instead of commas (",").
