### 2025-07-17 #8334 @Annchovy
A few comments on the paper: https://github.com/povilaskarvelis/e2p-simulator/issues/5#issue-3239402596
