### 2025-07-21 povilaskarvelis/e2p-simulator#3 @povilaskarvelis
In the first example, you set d to the wrong value 1.59, instead of 1.58, that's why the results don't match. These are analytical solutions, so they should work every time.

In the second case, we use data simulations and thus there is a small sampling error that leads to 0.01 value fluctuation in PR-AUC. If you toggle 'Get precise estimates', it should give 0.32 - I just increased it to 10 times more samples than in the default case to make sure the error is below 0.01.
