### 2025-07-23 povilaskarvelis/e2p-simulator#5 @povilaskarvelis
Thank you for these suggestions! The figure has been updated to provide a summary of the main functionality. The citation formatting has also been fixed.

We have also used this opportunity to reword a couple of paragraphs to be a bit more clear and more in line with the current version of the tool.
