### 2025-07-23 d55e52b
updates in response to paper review
