+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'E2P Simulator: An Interactive Tool for Translating Effect Sizes into Predictive Utility'
+title: 'E2P Simulator: An Interactive Tool for Estimating Real-World Predictive Utility of Research Findings'
 tags:
   - JavaScript
   - statistics
@@ -20,36 +20,35 @@ authors:
 affiliations:
   - name: Krembil Centre for Neuroinformatics, Centre for Addiction and Mental Health, University of Toronto, Canada
     index: 1
-date: 13 March 2024
+date: 22 July 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-The E2P Simulator (Effect-to-Prediction Simulator) is an interactive web-based tool designed to help researchers translate effect sizes into predictive utility. It visually and quantitatively demonstrates the relationship between commonly reported statistical metrics (such as Cohen's d and Pearson's r) and predictive metrics (such as AUC and PR-AUC) while accounting for measurement reliability and outcome base rates (\autoref{fig:interface}).
+E2P Simulator (Effect-to-Prediction Simulator) allows researchers to interactively and quantitatively explore the relationship between effect sizes (e.g., Cohen's d, Odds Ratio, Pearson's r), the corresponding predictive performance (e.g., ROC-AUC, Sensitivity, Specificity, Accuracy, etc.), and real-world predictive utility (e.g., PPV, NPV, PR-AUC, MCC, Net Benefit, etc.) by accounting for measurement reliability and outcome base rates. (\autoref{fig:interface}).
 
-![The E2P Simulator interface showing the relationship between effect size (Cohen's d and alternative effect size metrics) and predictive metrics. The left panel allows users to adjust parameters such as effect size, reliability, and base rate. The middle panel displays the resulting distributions with the adjustable threshold. The panel on the right displays Receiver Operating Charactersitic (ROC) and Precision-Recall (PR) curves with corresponding Area Under the Curve (AUC) metrics. The panel at the bottom displays the most common predictive metrics.\label{fig:interface}](interface_paper.png)
+![E2P Simulator interface with a high-level summary of its use. The left panel allows users to adjust parameters such as effect size, reliability, and base rate. The middle panel displays the resulting distributions with the adjustable threshold. The panel on the right displays Receiver Operating Charactersitic (ROC) and Precision-Recall (PR) curves with corresponding Area Under the Curve (AUC) metrics, as well as Decision Curve Analysis (DCA) plot. The panel at the bottom displays the most common predictive metrics.\label{fig:interface}](interface_paper.png)
 
-The E2P Simulator has several potential applications:
+E2P Simulator has several potential applications:
 
-1. **Interpretation of findings**: It helps researchers move beyond arbitrary "small/medium/large" effect size labels and ground the interpretation of their findings in terms of predictive value in specific contexts.
+1. **Interpretation of findings**: It helps researchers move beyond arbitrary "small/medium/large" effect size labels and misleading predictive metrics by grounding their interpretation in estimated real-world predictive utility. 
 
-2. **Research planning**: Being able to easily derive what effect sizes are needed to achieve the desired predictive performance allows researchers to plan their studies more effectively and allocate resources more efficiently.
+2. **Research planning**: Being able to easily derive what effect sizes and predictive performance are needed to achieve a desired real-world predictive performance allows researchers to plan their studies more effectively and allocate resources more efficiently.
 
-3. **Education**: The simulator's interactive design makes it a valuable teaching tool, helping students and researchers develop a more intuitive understanding of how different abstract statistical metrics relate to one another and to real-world impact.
+3. **Education**: The simulator's interactive design makes it a valuable teaching tool, helping researchers develop a more intuitive understanding of how different abstract statistical metrics relate to one another and to real-world utility.
 
[…]
