+++ paper/paper.bib
@@ -185,6 +185,15 @@
   doi={10.1080/00031305.2016.1154108}
 }
 
+@inproceedings{brabec2020model,
+  title={On model evaluation under non-constant class imbalance},
+  author={Brabec, Jan and Kom{\'a}rek, Tom{\'a}{\v{s}} and Franc, Vojt{\v{e}}ch and Machlica, Luk{\'a}{\v{s}}},
+  booktitle={International Conference on Computational Science},
+  pages={74--87},
+  year={2020},
+  organization={Springer}
+}
+
 @article{d3js,
   title={D3: Data-Driven Documents},
   author={Bostock, Michael and Ogievetsky, Vadim and Heer, Jeffrey},
