### 2025-07-24 225d9a6
fixed dois of new references
