+++ paper/paper.md
@@ -42,7 +42,7 @@ This tool has been designed with biomedical and behavioral sciences in mind, par
 
 # Statement of Need
 
-In biomedical and behavioral sciences, the focus on statistical significance, which quantifies the probability of the effect being real, often overshadows effect sizes, which quantify the practical significance of the effect [@wasserstein2016asa]. This emphasis, combined with a methodological disconnect between classical statistics and predictive modeling, frequently leads researchers to misinterpret any statistically significant finding as clinically meaningful, regardless of its effect size [@funder2019evaluating;@wasserstein2019moving]. This misinterpretation is particularly problematic in areas such as biomarker research and precision medicine/psychiatry, where the goal is to find robust predictors of treatment response or disease course for individual patients [@monsarrat2018intriguing;@abi2023candidate]. Furthermore, factors such as measurement reliability, which attenuates effect sizes [@karvelis2023individual;@karvelis2025clarifying], and outcome base rates, which limit predictive power in real-world contexts [@ozenne2015precision;@abi2016search;@large2018role], are often overlooked in evaluating both individual predictors and predictive models [brabec2020model], leading to unlrealistic expectations, ineffective research planning, and resource misalocation. 
+In biomedical and behavioral sciences, the focus on statistical significance, which quantifies the probability of the effect being real, often overshadows effect sizes, which quantify the practical significance of the effect [@wasserstein2016asa]. This emphasis, combined with a methodological disconnect between classical statistics and predictive modeling, frequently leads researchers to misinterpret any statistically significant finding as clinically meaningful, regardless of its effect size [@funder2019evaluating;@wasserstein2019moving]. This misinterpretation is particularly problematic in areas such as biomarker research and precision medicine/psychiatry, where the goal is to find robust predictors of treatment response or disease course for individual patients [@monsarrat2018intriguing;@abi2023candidate]. Furthermore, factors such as measurement reliability, which attenuates effect sizes [@karvelis2023individual;@karvelis2025clarifying], and outcome base rates, which limit predictive power in real-world contexts [baldessarini1983predictive;@ozenne2015precision;@abi2016search;@large2018role;brabec2020model], are often overlooked in evaluating both individual predictors and predictive models, leading to unlrealistic expectations, ineffective research planning, and resource misalocation. 
 
 E2P Simulator addressess these challenges by providing an interactive platform where researchers can explore the relationships among all of these factors. Similar to how GPower [@faul2007g] is used to explore the relationships between effect size, sample size and significance levels to perform *power analysis*, E2P Simulator can be used to explore the relationships between effect size, predictive performance, and real-world predictive utility to perform *predictive utility analysis*. By making the relationships between effect sizes, reliability, base rates, and predictive metrics explicit, E2P Simulator enables researchers to interpret findings more accurately, design more impactful studies, and communicate results more clearly to a broader audience.
 
