### 2025-07-24 fd2a0c3
fixed dois of new references
