+++ paper/paper.bib
@@ -191,7 +191,7 @@
   booktitle={International Conference on Computational Science},
   pages={74--87},
   year={2020},
-  organization={Springer} 
+  organization={Springer}, 
   doi={doi.org/10.1007/978-3-030-50423-6_6}
 }
 
@@ -203,7 +203,7 @@
   number={5},
   pages={569--573},
   year={1983},
-  publisher={American Medical Association}
+  publisher={American Medical Association},
   doi={10.1001/archpsyc.1983.01790050095011}
 }
 
