### 2025-08-01 3ff6d1f
added PR-AUC to multivariable simulators and updated related getting started sections
