+++ js/mahalanobis.js
@@ -1,476 +1,323 @@
 (function() {
-// Change to module-specific names to avoid conflict
-const mahalanobisWidth = 600;
-const mahalanobisHeight = 400;
-const mahalanobisMargin = { top: 20, right: 30, bottom: 40, left: 50 };
-
-// Clinical utility threshold (can be made adjustable later)
-const CLINICAL_UTILITY_THRESHOLD = 2.5;
-
 // Global variables to track datasets and colors
-let datasets = [];
+    let dDatasets = [], prAucDatasets = [];
 const colors = [
-    '#008080', // Teal (keep this as primary)
-    '#E63946', // Bright Red
-    '#1E88E5', // Bright Blue
-    '#FFA726', // Bright Orange
-    '#9C27B0', // Bright Purple
-    '#00A896', // Bright Teal
-    '#26A69A', // Medium Teal
-    '#7B1FA2'  // Deep Purple
-];
+        '#008080', '#E63946', '#FFA726', '#1E88E5', 
+        '#9C27B0', '#00A896', '#26A69A', '#7B1FA2'
+    ];
+    let nextColorIndex = 0;
 
-// Track active curve and datasets
-let activeCurve = null;
-let dChart = null;
-let nextColorIndex = 0; // Track which color to use next
+    // Track active curves and chart instances
+    let dActiveCurve, prAucActiveCurve;
+    let dChart = null, prAucChart = null;
 
-function initializeMahalanobisPlot() {
-    // Create SVG container
-    const svg = d3.select("#mahalanobis-plot")
-        .append("svg")
-        .attr("width", "100%")
-        .attr("height", "100%")
-        .attr("viewBox", `0 0 ${mahalanobisWidth} ${mahalanobisHeight}`)
-        .attr("preserveAspectRatio", "xMidYMid meet");
+    function initializeMahalanobis() {
+        const chartContainer = document.getElementById('chartContainer');
+        chartContainer.style.display = 'flex';
+        chartContainer.style.flexDirection = 'column';
+        chartContainer.style.gap = '20px';
+        
+        let legendContainer = chartContainer.querySelector('.chart-legend');
+        if (!legendContainer) {
+            legendContainer = document.createElement('div');
+            legendContainer.className = 'chart-legend';
+            legendContainer.style.textAlign = 'center';
+            legendContainer.style.marginTop = '10px';
+            chartContainer.insertBefore(legendContainer, chartContainer.firstChild);
+        }
+        
+        initializeDChart();
+        initializePrAucChart();
+        setupInputListeners();
+        updatePlot();
+    }
 
-    // Add axes
-    const xScale = d3.scaleLinear()
-        .domain([0, 10])  // Number of features
-        .range([mahalanobisMargin.left, mahalanobisWidth - mahalanobisMargin.right]);
-
-    const yScale = d3.scaleLinear()
-        .domain([0, 5])   // Mahalanobis distance
-        .range([mahalanobisHeight - mahalanobisMargin.bottom, mahalanobisMargin.top]);
-
-    // Add X axis
-    svg.append("g")
-        .attr("transform", `translate(0,${mahalanobisHeight - mahalanobisMargin.bottom})`)
-        .call(d3.axisBottom(xScale))
-        .append("text")
-        .attr("x", mahalanobisWidth / 2)
-        .attr("y", 35)
-        .attr("fill", "black")
-        .text("Number of Features");
-
-    // Add Y axis
-    svg.append("g")
-        .attr("transform", `translate(${mahalanobisMargin.left},0)`)
-        .call(d3.axisLeft(yScale))
-        .append("text")
-        .attr("transform", "rotate(-90)")
-        .attr("y", -35)
[…]
