### 2025-08-01 bf077e1
updated readme with the latest zenodo link after new release
