### 2025-08-08 1cff8a9
more predictive metrics + max J and F1 buttons
