+++ js/binary.js
@@ -75,6 +75,23 @@ function computeMetricsForBinaryDistributions(d, baseRate, sigma1, sigma2, thres
     const mcc = (TP * TN - FP * FN) / 
         Math.sqrt((TP + FP) * (TP + FN) * (TN + FP) * (TN + FN) || 1); // Avoid division by zero
     
+    // Calculate Likelihood Ratios
+    const lrPlus = sensitivity / (1 - specificity);
+    const lrMinus = (1 - sensitivity) / specificity;
+    const dor = lrPlus / lrMinus;
+    const youden = sensitivity + specificity - 1;
+    const gMean = Math.sqrt(sensitivity * specificity);
+    const nnd = 1 / (sensitivity + specificity - 1);
+    const nnm = 1 / ((1 - specificity) + (1 - sensitivity));
+
+    // Post-test probabilities
+    const preTestOdds = baseRate / (1 - baseRate);
+    const postTestOddsPlus = preTestOdds * lrPlus;
+    const postTestOddsMinus = preTestOdds * lrMinus;
+
+    const postTestProbPlus = postTestOddsPlus / (1 + postTestOddsPlus);
+    const postTestProbMinus = postTestOddsMinus / (1 + postTestOddsMinus);
+
     return {
         fpr: thresholdFPR,
         tpr: thresholdTPR,
@@ -85,9 +102,103 @@ function computeMetricsForBinaryDistributions(d, baseRate, sigma1, sigma2, thres
         accuracy,
         balancedAccuracy,
         f1Score,
-        mcc
+        mcc,
+        lrPlus,
+        lrMinus,
+        dor,
+        youden,
+        gMean,
+        nnd,
+        nnm,
+        postTestProbPlus,
+        postTestProbMinus
     };
 }
+
+// Helper function to find optimal threshold for Youden's J or F1
+function findOptimalThresholdBinary(metricType = 'youden') {
+    // Iterative hill-climbing search with shrinking step for very high precision
+    try {
+        const trueD = parseFloat(document.getElementById('true-difference-number-bin').value);
+        const icc1 = parseFloat(document.getElementById('icc1-slider').value);
+        const icc2 = parseFloat(document.getElementById('icc2-slider').value);
+        const kappa = parseFloat(document.getElementById('kappa-slider').value);
+        
+        // Compute the observed distributions
+        const obsD = trueD * Math.sqrt(Math.sin((Math.PI/2) * kappa));        
+        const ddif = (currentView === 'true') ? trueD : obsD;
+
+        const baseRate = parseFloat(document.getElementById('base-rate-slider').value) / 100;
+
+        const sigma1 = currentView === 'true' ? 1 : 1 / Math.sqrt(icc1);
+        const sigma2 = currentView === 'true' ? 1 : 1 / Math.sqrt(icc2);
+
+        // Helper to evaluate metric value at a given threshold
+        const metricAt = t => {
+            const m = computeMetricsForBinaryDistributions(ddif, baseRate, sigma1, sigma2, t);
+            return metricType === 'f1' ? m.f1Score : m.youden;
+        };
+
+        // Use ternary search for unimodal optimization (more robust than hill-climb)
+        let left = -8;
+        let right = 8;
+        const EPSILON = 1e-6;
+
+        while (right - left > EPSILON) {
+            const third = (right - left) / 3;
+            const m1 = left + third;
+            const m2 = right - third;
+            
+            if (metricAt(m1) < metricAt(m2)) {
+                left = m1;
+            } else {
+                right = m2;
+            }
+        }
+
+        const bestT = (left + right) / 2;
+        return parseFloat(bestT.toFixed(5));
+    } catch (err) {
[…]
