### 2025-08-13 fb6b3b4
added max J and F1 buttons to continuous + speed improvements
