+++ js/continuous.js
@@ -52,7 +52,7 @@ function generateLabeledData(r) {
     // Check the state of the precise estimates checkbox
     const preciseCheckbox = document.getElementById("precise-estimates-checkbox-cont");
     // Corrected numPoints logic from previous user interaction if needed
-    const numPoints = preciseCheckbox && preciseCheckbox.checked ? 400000 : 50000; 
+    const numPoints = preciseCheckbox && preciseCheckbox.checked ? 800000 : 10000; 
     const numPlotPoints = 4000; // Keep plot points lower for performance
     const meanX = 0, meanY = 0, stdDevX = 1, stdDevY = 1;
     const baseRate = parseFloat(document.getElementById("base-rate-slider-cont").value) / 100;
@@ -158,6 +158,15 @@ function computePredictiveMetrics(threshold, data) {
     const nnd = 1 / (sensitivity + specificity - 1);
     const nnm = 1 / ((1 - specificity) + (1 - sensitivity));
 
+    // Cohen's kappa (chance-corrected agreement) using observed marginals
+    const pYesTrue = data.filter(d => d.trueClass === 1).length / data.length;
+    const pYesPred = predictions.filter(p => p === 1).length / data.length;
+    const pNoTrue = 1 - pYesTrue;
+    const pNoPred = 1 - pYesPred;
+    const po = accuracy;
+    const peChance = pYesTrue * pYesPred + pNoTrue * pNoPred;
+    const kappa = (po - peChance) / (1 - peChance || 1);
+
     // Post-test probabilities
     const baseRate = data.filter(d => d.trueClass === 1).length / data.length;
     const preTestOdds = baseRate / (1 - baseRate);
@@ -186,8 +195,135 @@ function computePredictiveMetrics(threshold, data) {
         nnm,
         postTestProbPlus,
         postTestProbMinus,
+        kappa,
         fpr: 1 - specificity
     };
+}
+
+// Find optimal threshold on current simulated data for a given metric ('youden' or 'f1')
+function findOptimalThresholdContinuous(metricType = 'youden') {
+    try {
+        if (!currentLabeledData || currentLabeledData.length === 0) {
+            return thresholdValue;
+        }
+
+        // Build sweep arrays on a subsample to avoid heavy computation on very large datasets
+        const MAX_OPT_POINTS = isPreciseModeEnabled() ? 200000 : 50000; // higher cap in precise mode
+        const { xs, labels, posSuffix, negSuffix, posTotal, negTotal } = buildSweepData(currentLabeledData, MAX_OPT_POINTS);
+        const n = xs.length;
+        if (n === 0) return thresholdValue;
+
+        // Scan candidate thresholds at a controlled resolution
+        const MAX_CANDIDATES = isPreciseModeEnabled() ? 4000 : 2000;
+        const stepIdx = Math.max(1, Math.floor(n / MAX_CANDIDATES));
+
+        let bestMetric = -Infinity;
+        let bestThreshold = xs[0];
+
+        for (let i = 0; i < n; i += stepIdx) {
+            const TP = posSuffix[i];
+            const FP = negSuffix[i];
+            const FN = posTotal - TP;
+            const TN = negTotal - FP;
+
+            const sensitivity = TP / (TP + FN) || 0;
+            const specificity = TN / (TN + FP) || 0;
+            const ppv = TP / (TP + FP) || 0;
+            const youden = sensitivity + specificity - 1;
+            const f1 = 2 * (ppv * sensitivity) / (ppv + sensitivity) || 0;
+
+            const value = (metricType === 'f1') ? f1 : youden;
+            if (value > bestMetric) {
+                bestMetric = value;
+                bestThreshold = xs[i];
+            }
+        }
+
[…]
