+++ js/sample-size-continuous.js
@@ -1,0 +1,191 @@
+// Continuous outcome sample size calculator (Riley et al., BMJ 2020 + Stat Med 2019)
+// Computes n from three criteria and displays summary + simple charts
+
+(function(){
+    function zForCI(ci){
+        const level = parseFloat(ci);
+        if (level >= 0.999) return 3.29;
+        if (level >= 0.99) return 2.576;
+        if (level >= 0.95) return 1.96;
+        if (level >= 0.90) return 1.645;
+        return 1.96;
+    }
+
+    function val(id){
+        const el = document.getElementById(id);
+        if (!el) return null;
+        if (el.tagName === 'SELECT' || el.type === 'select-one') return el.value;
+        const v = parseFloat(el.value);
+        return isNaN(v) ? null : v;
+    }
+
+    function setHTML(id, text){
+        const el = document.getElementById(id);
+        if (el) el.textContent = text;
+    }
+
+    function formatInt(x){
+        if (!isFinite(x) || isNaN(x)) return '-';
+        return Math.ceil(Math.max(0, x)).toLocaleString();
+    }
+
+    // Criterion 1: target shrinkage via adjusted R2 approximation
+    // We approximate n ensuring adjusted R2 not too optimistic; closed-form here focuses on delta R2
+    function computeOptimismNR2(p, R2, delta){
+        // From rearranging adjusted R2 difference <= delta approx
+        // n >= p + 1 + p(1 - R2)/delta
+        return (p + 1) + (p * (1 - R2)) / Math.max(1e-6, delta);
+    }
+
+    // Criterion 2: mean(Y) precision (for intercept)
+    function computeMeanPrecisionN(sdY, margin, ci){
+        const z = zForCI(ci);
+        const m = Math.max(1e-6, margin);
+        return (z * sdY / m) * (z * sdY / m);
+    }
+
+    // Criterion 3: proxy shrinkage bound using same delta form but with smaller delta equivalent to (1-S)*(1-R2)
+    function computeShrinkageN(p, R2, S){
+        // Use heuristic: delta_S = (1 - S) * (1 - R2)
+        const deltaS = (1 - S) * Math.max(1e-6, (1 - R2));
+        return computeOptimismNR2(p, R2, Math.max(1e-6, deltaS));
+    }
+
+    function syncPair(sliderId, inputId){
+        const s = document.getElementById(sliderId);
+        const i = document.getElementById(inputId);
+        if (!s || !i) return;
+        s.addEventListener('input', () => { i.value = s.value; update(); });
+        i.addEventListener('input', () => { s.value = i.value; update(); });
+    }
+
+    function drawLineChart(canvasId, xs, ys, chartTitle, xAxisTitle, pVal){
+        const ctx = document.getElementById(canvasId);
+        if (!ctx || typeof Chart === 'undefined') return;
+        if (ctx._chart) { ctx._chart.destroy(); }
+
+        const xOptions = {
+            title: { display: true, text: xAxisTitle, font: { size: 18 } },
+            ticks: { font: { size: 14 } },
+            grid: { display: false, drawBorder: false }
+        };
+
+        if (pVal) {
+            xOptions.min = 0;
+            xOptions.max = pVal + 1;
+            xOptions.type = 'linear';
+            xOptions.ticks.callback = function(value) {
+                if (value <= 0 || value > pVal) return null;
+                if (Number.isInteger(value)) return value;
+                return null;
+            };
+
+            if (pVal <= 10) {
+                xOptions.ticks.stepSize = 1;
+            } else {
+                xOptions.ticks.maxTicksLimit = 10;
+            }
+        } else {
[…]
