+++ js/sample-size-binary.js
@@ -1,0 +1,231 @@
+// Binary outcome sample size calculator (Riley et al., BMJ 2020)
+// Computes n from three criteria and displays summary + simple charts
+
+(function(){
+    function zForCI(ci){
+        const level = parseFloat(ci);
+        if (level >= 0.999) return 3.29;
+        if (level >= 0.99) return 2.576;
+        if (level >= 0.95) return 1.96;
+        if (level >= 0.90) return 1.645;
+        return 1.96;
+    }
+
+    function clamp01(x){ return Math.max(0, Math.min(1, x)); }
+
+    function computeShrinkageN(p, S, r2cs){
+        const k = -Math.log(1 - r2cs);
+        if (!isFinite(k) || k <= 0) return NaN;
+        return p / ((1 - S) * k);
+    }
+
+    function computeOptimismN(p, delta, r2cs){
+        const k = -Math.log(1 - r2cs);
+        if (!isFinite(k) || k <= 0) return NaN;
+        return p / (delta * k);
+    }
+
+    // Removed mean risk precision criterion from calculations/plots
+
+    // Mean absolute prediction error criterion (B2) based on van Smeden et al. (2016)
+    // ln(MAPE) = -0.508 - 0.544 ln(n) + 0.259 ln(phi) + 0.504 ln(P)
+    // => n = exp(( -0.508 + 0.259 ln(phi) + 0.504 ln(P) - ln(MAPE) ) / 0.544)
+    function computeMapeN(p, targetMape, phi){
+        const P = Math.max(1, p);
+        const mape = Math.max(1e-6, targetMape);
+        // Ensure phi <= 0.5 by symmetry
+        let phiEff = Math.max(1e-6, Math.min(phi, 1 - phi));
+        // Only validated for P <= 30
+        if (P > 30) return NaN;
+        const num = -0.508 + 0.259 * Math.log(phiEff) + 0.504 * Math.log(P) - Math.log(mape);
+        const den = 0.544;
+        return Math.exp(num / den);
+    }
+
+    function formatInt(x){
+        if (!isFinite(x) || isNaN(x)) return '-';
+        return Math.ceil(Math.max(0, x)).toLocaleString();
+    }
+
+    function val(id){
+        const el = document.getElementById(id);
+        if (!el) return null;
+        if (el.tagName === 'SELECT' || el.type === 'select-one') return el.value;
+        const v = parseFloat(el.value);
+        return isNaN(v) ? null : v;
+    }
+
+    function setHTML(id, text){
+        const el = document.getElementById(id);
+        if (el) el.textContent = text;
+    }
+
+    function syncPair(sliderId, inputId){
+        const s = document.getElementById(sliderId);
+        const i = document.getElementById(inputId);
+        if (!s || !i) return;
+        s.addEventListener('input', () => { i.value = s.value; update(); });
+        i.addEventListener('input', () => { s.value = i.value; update(); });
+    }
+
+    function drawMultiLineChart(canvasId, xs, series, chartTitle, xAxisTitle, pVal){
+        const ctx = document.getElementById(canvasId);
+        if (!ctx || typeof Chart === 'undefined') return;
+        if (ctx._chart) { ctx._chart.destroy(); }
+        const plugins = [];
+        if (window.customLegendPlugin) plugins.push(window.customLegendPlugin);
+
+        const tickOptions = {
+            font: { size: 14 },
+            callback: function(value) {
+                if (value <= 0 || value > pVal) return null;
+                if (Number.isInteger(value)) return value;
+                return null;
+            }
+        };
+
+        if (pVal <= 10) {
+            tickOptions.stepSize = 1;
+        } else {
+            tickOptions.maxTicksLimit = 10;
+        }
+
[…]
