### 2025-09-05 f510237
fixed DCA! and updated sample size calculation for continuous
