+++ js/dca.js
@@ -1,5 +1,5 @@
 (function() {
-// DCA Module - Modular Decision Curve Analysis implementation
+// DCA Module - Simplified Decision Curve Analysis implementation
 const DCAModule = {
     // State variables for each DCA instance
     instances: new Map(),
@@ -8,8 +8,6 @@ const DCAModule = {
     init: function(instanceId, config) {
         this.instances.set(instanceId, {
             plotSelector: config.plotSelector,
-            thresholdMin: config.thresholdMin || 0.05,
-            thresholdMax: config.thresholdMax || 0.30,
             initialized: false,
             onThresholdChange: config.onThresholdChange || (() => {})
         });
@@ -44,7 +42,11 @@ const DCAModule = {
             const {
                 sensitivity,
                 specificity,
-                baseRate
+                baseRate,
+                // ROC curve data for proper DCA calculation
+                FPR,
+                TPR,
+                thresholds
             } = data;
             
             // Calculate net benefit across threshold probabilities using current model performance
@@ -58,10 +60,59 @@ const DCAModule = {
             const treatNoneBenefits = [];
             
             for (let pt = ptMin; pt <= ptMax; pt += step) {
-                // Use current threshold's sensitivity/specificity for all threshold probabilities
                 const odds = pt / (1 - pt);
                 
-                const netBenefit = (sensitivity * baseRate) - ((1 - specificity) * (1 - baseRate) * odds);
+                // Calculate sensitivity and specificity at threshold probability pt
+                let sensitivityAtPt, specificityAtPt;
+                
+                if (FPR && TPR && FPR.length > 0 && TPR.length > 0) {
+                    // Use ROC curve data to find sensitivity and specificity at threshold pt
+                    // In DCA, pt represents the probability threshold for classification
+                    // We need to interpolate sensitivity and specificity from the ROC curve
+                    
+                    // For DCA, we need to find the point on the ROC curve that corresponds
+                    // to using pt as the classification threshold
+                    // This is a complex mapping that depends on the underlying prediction model
+                    
+                    // As a first approximation, we'll use the relationship:
+                    // pt = 0 (treat everyone) -> FPR = 1, TPR = 1
+                    // pt = 1 (treat no one) -> FPR = 0, TPR = 0
+                    // Linear interpolation: targetFPR = 1 - pt
+                    const targetFPR = 1 - pt;
+                    
+                    // Find the two closest points for interpolation
+                    let lowerIndex = 0;
+                    let upperIndex = FPR.length - 1;
+                    
+                    for (let i = 0; i < FPR.length - 1; i++) {
+                        if (FPR[i] >= targetFPR && FPR[i + 1] <= targetFPR) {
+                            lowerIndex = i;
+                            upperIndex = i + 1;
+                            break;
+                        }
+                    }
+                    
+                    // Linear interpolation
+                    const fpr1 = FPR[lowerIndex];
+                    const fpr2 = FPR[upperIndex];
+                    const tpr1 = TPR[lowerIndex];
+                    const tpr2 = TPR[upperIndex];
+                    
+                    if (fpr1 === fpr2) {
+                        sensitivityAtPt = tpr1;
+                        specificityAtPt = 1 - fpr1;
+                    } else {
+                        const weight = (targetFPR - fpr1) / (fpr2 - fpr1);
+                        sensitivityAtPt = tpr1 + weight * (tpr2 - tpr1);
+                        specificityAtPt = 1 - targetFPR;
+                    }
+                } else {
+                    // Fallback to provided sensitivity/specificity if ROC data not available
+                    sensitivityAtPt = sensitivity;
+                    specificityAtPt = specificity;
+                }
+                
[…]
