+++ js/sample-size-continuous.js
@@ -30,26 +30,23 @@
     }
 
     // Criterion 1: target shrinkage via adjusted R2 approximation
-    // We approximate n ensuring adjusted R2 not too optimistic; closed-form here focuses on delta R2
-    function computeOptimismNR2(p, R2, delta){
-        // From rearranging adjusted R2 difference <= delta approx
-        // n >= p + 1 + p(1 - R2)/delta
-        return (p + 1) + (p * (1 - R2)) / Math.max(1e-6, delta);
-    }
-
-    // Criterion 2: mean(Y) precision (for intercept)
-    function computeMeanPrecisionN(sdY, margin, ci){
-        const z = zForCI(ci);
-        const m = Math.max(1e-6, margin);
-        return (z * sdY / m) * (z * sdY / m);
-    }
-
-    // Criterion 3: proxy shrinkage bound using same delta form but with smaller delta equivalent to (1-S)*(1-R2)
     function computeShrinkageN(p, R2, S){
-        // Use heuristic: delta_S = (1 - S) * (1 - R2)
-        const deltaS = (1 - S) * Math.max(1e-6, (1 - R2));
-        return computeOptimismNR2(p, R2, Math.max(1e-6, deltaS));
+        // Riley et al. formula: n = p / ((S-1) × [ln(1-R²/S)])
+        const denominator = (S - 1) * Math.log(1 - R2 / S);
+        if (!isFinite(denominator) || denominator <= 0) return NaN;
+        return p / denominator;
     }
+
+    // Criterion 2: residual SD precision (N = 234 + p for ≤10% multiplicative error)
+    function computeResidualSDN(p){
+        return 234 + p;
+    }
+
+    // Criterion 3: optimism-based correction (n ≥ 1 + (p * (1 - R²)) / δ)
+    function computeOptimismN(p, R, delta){
+        return 1 + (p * (1 - R * R)) / delta;
+    }
+
 
     function syncPair(sliderId, inputId){
         const s = document.getElementById(sliderId);
@@ -59,53 +56,33 @@
         i.addEventListener('input', () => { s.value = i.value; update(); });
     }
 
-    function drawLineChart(canvasId, xs, ys, chartTitle, xAxisTitle, pVal){
+    function drawMultiLineChart(canvasId, xs, series, chartTitle, xAxisTitle, pVal){
         const ctx = document.getElementById(canvasId);
         if (!ctx || typeof Chart === 'undefined') return;
         if (ctx._chart) { ctx._chart.destroy(); }
+        const plugins = [];
+        if (window.customLegendPlugin) plugins.push(window.customLegendPlugin);
 
-        const xOptions = {
-            title: { display: true, text: xAxisTitle, font: { size: 18 } },
-            ticks: { font: { size: 14 } },
-            grid: { display: false, drawBorder: false }
-        };
-
-        if (pVal) {
-            xOptions.min = 0;
-            xOptions.max = pVal + 1;
-            xOptions.type = 'linear';
-            xOptions.ticks.callback = function(value) {
+        const tickOptions = {
+            font: { size: 14 },
+            callback: function(value) {
                 if (value <= 0 || value > pVal) return null;
                 if (Number.isInteger(value)) return value;
                 return null;
-            };
-
-            if (pVal <= 10) {
-                xOptions.ticks.stepSize = 1;
-            } else {
-                xOptions.ticks.maxTicksLimit = 10;
             }
+        };
+
+        if (pVal <= 10) {
+            tickOptions.stepSize = 1;
         } else {
-            xOptions.ticks.maxTicksLimit = 6;
+            tickOptions.maxTicksLimit = 10;
         }
 
-        const plugins = [];
-        if (window.customLegendPlugin) plugins.push(window.customLegendPlugin);
         ctx._chart = new Chart(ctx, {
             type: 'line',
[…]

[…]
