### 2025-09-05 4211d8b
fixed dca even more; updated instructions too
