+++ js/dca.js
@@ -49,9 +49,18 @@ const DCAModule = {
                 thresholds
             } = data;
             
+            // Debug: log the current threshold and metrics
+            console.log('DCA Debug:', {
+                currentThreshold: data.currentThreshold,
+                currentMetrics: data.currentMetrics ? {
+                    sensitivity: data.currentMetrics.sensitivity,
+                    specificity: data.currentMetrics.specificity
+                } : 'undefined'
+            });
+            
             // Calculate net benefit across threshold probabilities using current model performance
             const ptMin = 0.001;
-            const ptMax = 0.999;
+            const ptMax = 0.95;  // Further reduce upper limit for smoother curves
             const step = 0.001;
             
             const thresholdProbs = [];
@@ -62,65 +71,42 @@ const DCAModule = {
             for (let pt = ptMin; pt <= ptMax; pt += step) {
                 const odds = pt / (1 - pt);
                 
-                // Calculate sensitivity and specificity at threshold probability pt
-                let sensitivityAtPt, specificityAtPt;
+                // Calculate optimal sensitivity and specificity for this pt
+                let bestNetBenefit = -Infinity;
+                let sensitivityAtPt = 0;
+                let specificityAtPt = 0;
                 
                 if (FPR && TPR && FPR.length > 0 && TPR.length > 0) {
-                    // Use ROC curve data to find sensitivity and specificity at threshold pt
-                    // In DCA, pt represents the probability threshold for classification
-                    // We need to interpolate sensitivity and specificity from the ROC curve
-                    
-                    // For DCA, we need to find the point on the ROC curve that corresponds
-                    // to using pt as the classification threshold
-                    // This is a complex mapping that depends on the underlying prediction model
-                    
-                    // As a first approximation, we'll use the relationship:
-                    // pt = 0 (treat everyone) -> FPR = 1, TPR = 1
-                    // pt = 1 (treat no one) -> FPR = 0, TPR = 0
-                    // Linear interpolation: targetFPR = 1 - pt
-                    const targetFPR = 1 - pt;
-                    
-                    // Find the two closest points for interpolation
-                    let lowerIndex = 0;
-                    let upperIndex = FPR.length - 1;
-                    
-                    for (let i = 0; i < FPR.length - 1; i++) {
-                        if (FPR[i] >= targetFPR && FPR[i + 1] <= targetFPR) {
-                            lowerIndex = i;
-                            upperIndex = i + 1;
-                            break;
+                    // For each point on the ROC curve, calculate net benefit for this pt
+                    // Choose the point that gives the highest net benefit
+                    for (let i = 0; i < FPR.length; i++) {
+                        const currentSensitivity = TPR[i];
+                        const currentSpecificity = 1 - FPR[i];
+                        
+                        // Calculate net benefit for this ROC point at this pt
+                        const netBenefit = (currentSensitivity * baseRate) - ((1 - currentSpecificity) * (1 - baseRate) * odds);
+                        
+                        // Keep the ROC point that gives the highest net benefit
+                        if (netBenefit > bestNetBenefit) {
+                            bestNetBenefit = netBenefit;
+                            sensitivityAtPt = currentSensitivity;
+                            specificityAtPt = currentSpecificity;
                         }
                     }
-                    
-                    // Linear interpolation
-                    const fpr1 = FPR[lowerIndex];
-                    const fpr2 = FPR[upperIndex];
-                    const tpr1 = TPR[lowerIndex];
-                    const tpr2 = TPR[upperIndex];
-                    
-                    if (fpr1 === fpr2) {
[…]
