### 2025-09-11 c343a99
updated get ready with ss details
