### 2025-09-12 4f6b13c
misc small updates of the interface; U3 equation added
