### 2025-09-17 a07ee15
minor paper figure update
