### 2025-09-17 bf7a0b2
Add version selector with Netlify branch deployment support
