### 2025-09-17 36c634c
Update version selector to use Netlify branch deployments
