### 2025-09-17 ebfbf0a
Fix version selector URL format - remove extra v prefix
