### 2025-09-17 6c2fd10
Remove version selector (requires paid Netlify plan)
