### 2025-09-23 fe2fffe
get started updated with more detailed explanations
