### 2025-09-28 8c30a20
updated the examples
