+++ js/dca.js
@@ -11,6 +11,8 @@ const DCAModule = {
             initialized: false,
             onThresholdChange: config.onThresholdChange || (() => {})
         });
+        // Reset smoothing values for clean start
+        this.lastPtValue = undefined;
     },
     
     // Plot DCA for a given instance
@@ -59,9 +61,11 @@ const DCAModule = {
             });
             
             // Calculate net benefit across threshold probabilities using current model performance
+            // Use higher precision when precise estimates are enabled
+            const usePrecise = data.usePreciseEstimates || false;
             const ptMin = 0.001;
             const ptMax = 0.95;  // Further reduce upper limit for smoother curves
-            const step = 0.001;
+            const step = usePrecise ? 0.0001 : 0.001; // Much finer step when precise
             
             const thresholdProbs = [];
             const netBenefits = [];
@@ -134,55 +138,123 @@ const DCAModule = {
                 const currentThreshold = data.currentThreshold;
                 
                 // For the red marker, we need to find which pt value corresponds to the current classification threshold
-                // We do this by finding the pt that gives us the current sensitivity/specificity
-                let closestIndex = 0;
-                let minDiff = Infinity;
-            
-                // Find the pt that corresponds to the current threshold's actual net benefit
-                // Use a smoother approach to reduce marker jumping
-                let bestPt = 0.5;
-                let bestNetBenefit = 0;
-            
-            for (let i = 0; i < thresholdProbs.length; i++) {
-                const pt = thresholdProbs[i];
-                    const odds = pt / (1 - pt);
-                    
-                    // Calculate what net benefit the current threshold would give at this pt
-                    const currentThresholdNetBenefit = (currentMetrics.sensitivity * baseRate) - ((1 - currentMetrics.specificity) * (1 - baseRate) * odds);
-                    
-                    // Find the pt where this net benefit matches the DCA curve
-                    const diff = Math.abs(currentThresholdNetBenefit - netBenefits[i]);
-                    
-                    if (diff < minDiff) {
-                        minDiff = diff;
-                        closestIndex = i;
-                        bestPt = pt;
-                        bestNetBenefit = currentThresholdNetBenefit;
-                    }
-                }
+                // Use a more direct approach: find the pt where the current sensitivity/specificity would be optimal
                 
-                // Use interpolation to smooth the pt value
-                if (closestIndex > 0 && closestIndex < thresholdProbs.length - 1) {
-                    const prevPt = thresholdProbs[closestIndex - 1];
-                    const nextPt = thresholdProbs[closestIndex + 1];
-                    const prevNB = netBenefits[closestIndex - 1];
-                    const nextNB = netBenefits[closestIndex + 1];
-                    const currentNB = netBenefits[closestIndex];
+                // Calculate the current threshold's net benefit for each pt value
+                let bestPt = 0.5;
+                let bestDelta = Infinity;
+                let closestIndex = 0;
+                
+                // More stable approach: use direct mathematical calculation
+                // The pt value where current sensitivity/specificity would be optimal can be calculated directly
+                
+                if (FPR && TPR && FPR.length > 0 && TPR.length > 0) {
+                    // Find the closest ROC point to current sensitivity/specificity
+                    let rocIndex = 0;
+                    let minROCDist = Infinity;
                     
-                    // Linear interpolation to get a smoother pt value
-                    if (Math.abs(bestNetBenefit - prevNB) < Math.abs(bestNetBenefit - nextNB)) {
[…]
