### 2025-10-18 a6bf4d1
load get ready seperately; paper affiliations, refs and details polished
