+++ paper/paper.bib
@@ -34,12 +34,15 @@
 }
 
 
-@misc{karvelis2025clarifying,
-  title={Clarifying the reliability paradox: poor test-retest reliability attenuates group differences},
-  author={Karvelis, Povilas and Diaconescu, Andreea O},
-  year={2025},
-  publisher={PsyArXiv},
-  doi={10.31234/osf.io/z4yqe_v4}
+@article{karvelis2025clarifying,
+  title = {Clarifying the reliability paradox: Poor measurement reliability attenuates group differences},
+  author = {Karvelis, Povilas and Diaconescu, Andreea O.},
+  journal = {Frontiers in Psychology},
+  volume = {16},
+  number = {1592658},
+  year = {2025},
+  publisher = {Frontiers Media SA},
+  doi = {10.3389/fpsyg.2025.1592658}
 }
 
 @article{abi2016search,
