+++ paper/paper.md
@@ -16,17 +16,23 @@ authors:
     affiliation: 1
   - name: Andreea O. Diaconescu
     orcid: 0000-0002-3633-9757
-    affiliation: 1
+    affiliation: [1, 2, 3, 4]
 affiliations:
   - name: Krembil Centre for Neuroinformatics, Centre for Addiction and Mental Health, University of Toronto, Canada
     index: 1
-date: 22 July 2025
+  - name: Department of Psychiatry, University of Toronto, Toronto, ON, Canada
+    index: 2
+  - name: Institute of Medical Sciences, University of Toronto, Toronto, ON, Canada
+    index: 3
+  - name: Department of Psychology, University of Toronto, Toronto, ON, Canada
+    index: 4
+date: 18 October 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-E2P Simulator (Effect-to-Prediction Simulator) allows researchers to interactively and quantitatively explore the relationship between effect sizes (e.g., Cohen's d, Odds Ratio, Pearson's r), the corresponding predictive performance (e.g., ROC-AUC, Sensitivity, Specificity, Accuracy, etc.), and real-world predictive utility (e.g., PPV, NPV, PR-AUC, MCC, Net Benefit, etc.) by accounting for measurement reliability and outcome base rates. (\autoref{fig:interface}).
+E2P Simulator (Effect-to-Prediction Simulator) allows researchers to interactively and quantitatively explore the relationship between effect sizes (e.g., Cohen's d, Odds Ratio, Pearson's r), their discriminative ability (e.g., ROC-AUC, Sensitivity, Specificity, etc.), and real-world predictive value and clinical utility (e.g., PPV, NPV, PR-AUC, Net Benefit, etc.), while accounting for measurement reliability and outcome base rates (\autoref{fig:interface}).
 
 ![E2P Simulator interface with a high-level summary of its use. The left panel allows users to adjust parameters such as effect size, reliability, and base rate. The middle panel displays the resulting distributions with the adjustable threshold. The panel on the right displays Receiver Operating Charactersitic (ROC) and Precision-Recall (PR) curves with corresponding Area Under the Curve (AUC) metrics, as well as Decision Curve Analysis (DCA) plot. The panel at the bottom displays the most common predictive metrics.\label{fig:interface}](interface_paper.png)
 
@@ -38,13 +44,13 @@ E2P Simulator has several potential applications:
 
 3. **Education**: The simulator's interactive design makes it a valuable teaching tool, helping researchers develop a more intuitive understanding of how different abstract statistical metrics relate to one another and to real-world utility.
 
-This tool has been designed with biomedical and behavioral sciences in mind, particularly areas such as biomarker research and precision medicine/psychiatry. However, it may be just as useful for any area of research that focuses on personalization of interventions, such as within education and sports sciences.
+This tool has been designed with biomedical and behavioral sciences in mind, particularly areas such as biomarker research and precision medicine/psychiatry. However, it may be just as useful for any area of research that can is focused on predictive modelling and personalization, such as within forensic, education, and sports sciences.
 
 # Statement of Need
 
[…]
