### 2025-10-20 #8334 @povilaskarvelis
@blackdotbug - thank you for the comments!

We used different approaches for slightly different purposes: simple join for static elements (SVG containers), join with enter/update/exit functions for dynamic data with transitions (scatter points), and traditional enter-update-exit for complex nested elements (legends with editable content).
