### 2025-10-21 #8334 @blackdotbug
@jromanowska my checklist is complete. @povilaskarvelis thanks for the explanation on the d3 conventions! that makes total sense. Also, I'm happy to submit a pr for the responsiveness if that's something you want to pursue and would welcome the assist for. It's a stellar implementation of plotly and d3!
