### 2025-10-22 #8334 @jromanowska
@povilaskarvelis , could you go through the tasks marked "Additional Author Tasks" above?
