### 2025-10-22 #8334 @jromanowska
@povilaskarvelis - I think the affiliations should be re-formatted, look [at this example](https://joss.readthedocs.io/en/latest/paper.html#affiliations)

and also, could you fix the invalid DOI, as mentioned by the bot above?
