### 2025-10-22 359e951
fixed affiliations and doi formatting
