+++ paper/paper.bib
@@ -195,7 +195,7 @@
   pages={74--87},
   year={2020},
   organization={Springer}, 
-  doi={doi.org/10.1007/978-3-030-50423-6_6}
+  doi={10.1007/978-3-030-50423-6_6}
 }
 
 @article{baldessarini1983predictive,
