### 2025-10-22 #8334 @povilaskarvelis
@jromanowska I have fixed the issues and pushed the changes - hopefully it all works now
