### 2025-10-22 #8334 @povilaskarvelis
Zenodo doi of the latest release:

Povilas Karvelis. (2025). povilaskarvelis/e2p-simulator: E2P Simulator v1.2.0 (v1.2.0). Zenodo. https://doi.org/10.5281/zenodo.17112626
