### 2025-10-22 c48dad7
minor paper edits
