### 2025-10-22 #8334 @povilaskarvelis
@jromanowska sorry, I meant I would need to create new doi, not necessarily new release
