### 2025-10-22 abf2efa
minor affiliation fix
