+++ js/calibration.js
@@ -1,0 +1,1185 @@
+(function() {
+// Calibration Module for E2P Simulator
+// Explores calibration drift from test set to deployment
+
+// ============================================================================
+// Constants and Configuration
+// ============================================================================
+
+const CONFIG = {
+    // All calculations are now analytical - no sampling used
+    defaultTestEffectSize: 2.5,
+    defaultTestICC1: 0.7,
+    defaultTestICC2: 0.7,
+    defaultTestKappa: 1.0,
+    defaultTestBaseRate: 0.5,
+    defaultDeploymentEffectSize: 2.5,
+    defaultDeploymentICC1: 0.4,
+    defaultDeploymentICC2: 0.4,
+    defaultDeploymentKappa: 1.0,
+    defaultDeploymentBaseRate: 0.3
+};
+
+// ============================================================================
+// State Variables
+// ============================================================================
+
+let state = {
+    // Test set parameters
+    testEffectSize: CONFIG.defaultTestEffectSize,
+    testICC1: CONFIG.defaultTestICC1,
+    testICC2: CONFIG.defaultTestICC2,
+    testKappa: CONFIG.defaultTestKappa,
+    testBaseRate: CONFIG.defaultTestBaseRate,
+    
+    // Deployment set parameters
+    deploymentEffectSize: CONFIG.defaultDeploymentEffectSize,
+    deploymentICC1: CONFIG.defaultDeploymentICC1,
+    deploymentICC2: CONFIG.defaultDeploymentICC2,
+    deploymentKappa: CONFIG.defaultDeploymentKappa,
+    deploymentBaseRate: CONFIG.defaultDeploymentBaseRate,
+    
+    // Model parameters (fitted on test set)
+    modelBeta0: null,
+    modelBeta1: null,
+    
+    // Threshold value for calibration exploration
+    thresholdValue: 0,
+    
+    // Data
+    testData: null,
+    deploymentData: null,
+    calibrationData: null
+};
+
+// ============================================================================
+// Utility Functions
+// ============================================================================
+
+// Generate random normal samples
+function randomNormal(mean, stdDev, size) {
+    const samples = [];
+    for (let i = 0; i < size; i++) {
+        // Box-Muller transform
+        const u1 = Math.random();
+        const u2 = Math.random();
+        const z = Math.sqrt(-2 * Math.log(u1)) * Math.cos(2 * Math.PI * u2);
+        samples.push(mean + z * stdDev);
+    }
+    return samples;
+}
+
+// Normal CDF (using error function approximation)
+function normalCDF(x, mean = 0, stdDev = 1) {
+    const z = (x - mean) / stdDev;
+    const t = 1 / (1 + 0.2316419 * Math.abs(z));
+    const d = 0.3989423 * Math.exp(-z * z / 2);
+    const p = d * t * (0.3193815 + t * (-0.3565638 + t * (1.781478 + t * (-1.821256 + t * 1.330274))));
+    return z > 0 ? 1 - p : p;
+}
+
+// Normal PDF
+function normalPDF(x, mean = 0, stdDev = 1) {
+    const z = (x - mean) / stdDev;
+    return (1 / (stdDev * Math.sqrt(2 * Math.PI))) * Math.exp(-0.5 * z * z);
+}
+
+// Logistic function
+function logistic(x) {
+    return 1 / (1 + Math.exp(-x));
+}
+
+// Apply measurement reliability to observations
+function applyReliability(trueValues, reliability) {
+    const errorVariance = (1 - reliability);
+    return trueValues.map(trueVal => {
+        const error = randomNormal(0, Math.sqrt(errorVariance), 1)[0];
+        return trueVal + error;
+    });
+}
+
+// Wilson score confidence interval for proportions
+function wilsonInterval(successes, total, confidence = 0.95) {
[…]
