### 2025-10-24 4941837
readme citations updated
