### 2025-10-24 cad463c
readme citations updated
