+++ README.md
@@ -1,5 +1,5 @@
 # E2P Simulator
-[![DOI](https://zenodo.org/badge/934366440.svg)](https://doi.org/10.5281/zenodo.17112626)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17112626.svg)](https://doi.org/10.5281/zenodo.17112626)
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.08334/status.svg)](https://doi.org/10.21105/joss.08334)
 
 E2P Simulator (Effect-to-Prediction Simulator) allows researchers to interactively and quantitatively explore the relationship between effect sizes (e.g., Cohen's d, Odds Ratio, Pearson's r), the corresponding predictive performance (e.g., ROC-AUC, Sensitivity, Specificity, Accuracy, etc.), as well as real-world predictive value and clinical utility (e.g., PPV, NPV, PR-AUC, Net Benefit, etc.) while explicitly accounting for measurement reliability and outcome base rates.
